+++ Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "simple_auth_bot"
+name = "valine_bot"
 version = "0.1.0"
 edition = "2021"
