+++ src/features/logging.rs
@@ -1,0 +1,8 @@
+use serenity::{all::EventHandler, async_trait};
+
+pub struct Handler;
+
+
+#[async_trait]
+impl EventHandler for Handler {
+}
