+++ src/features/mod.rs
@@ -1,0 +1,5 @@
+mod auth;
+mod logging;
+
+pub use auth::Handler as AuthHandler;
+pub use logging::Handler as LoggingHandler;
